MLP LayerNorm dtype, agent.update() logging fix

+++ tdmpc2_jax/networks/mlp.py
@@ -24,7 +24,7 @@ class NormedLinear(nn.Module):
                  dtype=self.dtype,
                  param_dtype=self.param_dtype)(x)
 
-    x = self.norm()(x)
+    x = self.norm(dtype=self.dtype)(x)
 
     if self.activation is not None:
       x = self.activation(x)
